refactor: join reset button layout and handler in one file
